requirements.txt
arrow==0.10.0
click==6.7
contextlib2==0.5.5
croniter==0.3.16
dill==0.2.5
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.6
django-dynamic-preferences==1.1
docopt==0.4.0
Flask==0.12
future==0.16.0
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
persisting-theory==0.2.1
psycopg2==2.7.1
PyFunctional==1.0.0
python-dateutil==2.6.0
python-http-client==2.2.1
pytz==2017.2
requests==2.13.0
sendgrid==4.0.0
six==1.10.0
static3==0.7.0
tabulate==0.7.7
Werkzeug==0.12.1
celery==4.0.2
django-celery-results==1.0.1
